+++ etc/scrum-poker-performance/src/test/java/de/codescape/jira/plugins/scrumpoker/pages/ScrumPokerPage.java
@@ -1,5 +1,6 @@
 package de.codescape.jira.plugins.scrumpoker.pages;
 
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
@@ -23,7 +24,7 @@ public class ScrumPokerPage extends Page {
 
     @Override
     public boolean verifyPage() {
-        return issueLink.isDisplayed() && issueLink.getText().equals(issueKey);
+        return webDriver.findElement(By.className("scrum-poker-session")).isDisplayed();
     }
 
 }