+++ docs/changelog.md
@@ -15,6 +15,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: migrate old votes with question mark to new string format
 * test: improve test coverage and suppress unused warnings
 * test: include performance test suite for Jira Data Center
+* test: add markers to session list and detail view to identify pages
 
 ### [3.12.0] - 2019-04-05